--story=868374287 【Dashboard】增加最近访问的4个仓库 - 后端接口

+++ src/backend/ci/ext/tencent/stream/api-stream-tencent/src/main/kotlin/com/tencent/devops/stream/api/user/UserGitCIProjectResource.kt
@@ -89,6 +89,6 @@ interface UserGitCIProjectResource {
         userId: String,
         @ApiParam("多少条记录", required = false, defaultValue = "4")
         @QueryParam("pageSize")
-        size: Int?,
+        size: Long?
     ): Result<List<ProjectCIInfo>>
 }
